Account for more than one error on the same line.
TODO:
When a latex error occurs far away from the beginning of a line, as

This is a long line which is quite long really and here is the $\errorthingie$

Then the corresponding line in the log file might look like

l.26 .... here is the $\errorthingie$.

Which causes a faulty calculation of the column number. A patter search
approach might therefore be much more robust at column number calculations.


git-svn-id: https://vim-latex.svn.sourceforge.net/svnroot/vim-latex/trunk@175 a797aaa2-510f-0410-87df-a343b97812ff