Changing encoding from one locale to the next causes huge ugly screen flickers. Therefore only change encoding if the original encoding is not latin1. This will at least cause it to be smoother on machines using latin1 encoding.
The problem ofcourse will still be there for other locales. Therefore, look
for a more long term solution which does not require the encoding to be
changed back and forth with each imap execution.

If only someone had told me at the beginning not to use ascii > 127
characters!! I think in the next major rev, we will have to undertake a
major overhaul of all the macros to not use any of the higher ascii
characters. Maybe modify IMAP() to allow for multiple character placeholder
characters... (maybe '{+' and '+}' like in cream).

This will instantly remove all the internationalization and localization
problems.


git-svn-id: https://vim-latex.svn.sourceforge.net/svnroot/vim-latex/trunk@206 a797aaa2-510f-0410-87df-a343b97812ff