This is a merge from the b-newimaps branch. The branch was started when a need to robustify imaps.vim was felt. Before, it had a big dependence on "funky" characters. Since imaps.vim affected a lot of files, this commit has a lot of modified files.
Briefly, it changes the <<, >> and the <a-umlaut> characters throughout
latex-suite to 'unfunky' characters.
<< --- <+
>> --- +>
<a-umlaut> --- <++>

The syntax of IMAP and IMAP_PutTextWithMovement was also changed to accept
2 new optional arguments which specify the placeholder characters in the
a:rhs. This allows scriptwriters to generate long term maps which will not
depend on the internals of imaps.vim

The b-newimaps tip at the time of this merge has been tagged as
b-newimaps-merge-sa-1


git-svn-id: https://vim-latex.svn.sourceforge.net/svnroot/vim-latex/trunk@234 a797aaa2-510f-0410-87df-a343b97812ff